refactor(utils): 유틸 디렉토리 내의 파일 명 변경
