+++ src/utils/date.ts
@@ -1,7 +1,5 @@
-const getLocalDateString = () => {
+export const getLocalDateString = () => {
   const localDateTime = new Date().toISOString();
 
   return localDateTime.replace(/T/, '-').replace(/\..+/, '');
 };
-
-export default getLocalDateString;
