+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-export { default as pxToRem } from './pxToRem';
-export { default as createBreakPoint } from './createBreakPoint';
-export { default as getRandomNumber } from './getRandomNumber';
-export { default as getLocalDateString } from './getLocalDateString';
+export { pxToRem, createBreakPoint } from './style';
+export { getRandomNumberWithRange } from './random';
+export { getLocalDateString } from './date';
