+++ src/utils/random.ts
@@ -1,4 +1,4 @@
-const getRandomNumber = (min: number, max: number) => {
+export const getRandomNumberWithRange = (min: number, max: number) => {
   if (!Number.isInteger(min) || !Number.isInteger(max)) {
     throw new Error('정수만 입력할 수 있습니다.');
   }
@@ -9,5 +9,3 @@ const getRandomNumber = (min: number, max: number) => {
 
   return Math.floor(Math.random() * (max - min + 1)) + min;
 };
-
-export default getRandomNumber;
